Add utils and move consts and CreateTensorleapChartValues
